Create Rocket_DeltaVAllocation_GLOW_Algorithum
MATLAB code which can be used for designing two-stage rocket with inputs of required delta-V and payload capacity for a Highly Elliptical Orbit (HEO) around Earth. This algorithm minimizes the mass of the first stage of a two-stage rocket which also outputs delta-V allocation fractions, a plot of Initial Mass vs. Delta-V fraction and total mass of propellant required.